+++ chemodurov.artem/S5/create-args-map.hpp
@@ -6,7 +6,7 @@
 namespace chemodurov
 {
   template< typename F >
-  using ptr_t = F (Map< int, std::string >::*)(F f);
+  using ptr_t = F (Map< long, std::string >::*)(F f);
   using args_dic_t = Map< std::string, ptr_t< Printer > >;
   args_dic_t createArgsMap();
 }
